working draft of edit company
